+++ plugins/stage.lua
@@ -1,0 +1,22 @@
+-- Stage
+
+Stage = {}
+
+local abstraction_data = setmetatable({}, {__mode = "k"})
+
+
+
+-- ========== Enums ==========
+
+
+
+
+
+-- ========== Static Methods ==========
+
+
+
+
+
+-- ========== Metatables ==========
+