+++ plugins/main.lua
@@ -19,12 +19,13 @@ Classes = {
     "Helper",
     "Initialize",
     "Instance",
-    -- "Interactable",
+    "Interactable",
     "Interactable_Card",
     "Item",
     "Language",
     "List",
     "Mod",
+    "Monster_Card",
     "Net",
     "Object",
     "Player",
@@ -58,7 +59,7 @@ function __initialize()
     Buff.__initialize()
     Equipment.__initialize()
     Instance.__initialize()
-    -- Interactable.__initialize()
+    Interactable.__initialize()
     Item.__initialize()
 end