+++ CHANGELOG.md
@@ -5,6 +5,8 @@
     * Interactable_Card.new
     * Interactable_Card.find
     * Interactable_Card.wrap
+* Added Stage class
+    * 
 
 ### v1.1.9
 * Object : `create` now defaults to position (0, 0) if none is specified.