+++ CHANGELOG.md
@@ -17,8 +17,6 @@
     * Stage.new
     * Stage.find
     * Stage.wrap
-    * stage:set_log_icon
-    * stage:set_log_hidden
     * stage:set_index
     * stage:add_room
     * stage:get_room
@@ -28,6 +26,9 @@
     * stage:add_monster
     * stage:add_monster_loop
     * stage:clear_monsters
+    * stage:set_log_icon
+    * stage:set_log_view_start
+    * stage:set_log_hidden
 
 ### v1.1.9
 * Object : `create` now defaults to position (0, 0) if none is specified.